Corrected script names
