--- run_integration_tests.yaml
@@ -1,37 +1,0 @@
-name: Run unit tests
-
-on: [ push ]
-
-jobs:
-  unit_tests:
-    name: Run unit tests test on ${{ matrix.config.python-version }}
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: [3.8]
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
-      - uses: abatilo/actions-poetry@v2.0.0
-        with:
-          poetry-version: 1.1.4
-      - name: Install Poetry
-        run: poetry install
-      - name: Build Poetry
-        run: poetry build
-      - name: Install project whl
-        run: pip install dist/*.whl
-      - name: Install Lua dependencies
-        run: |
-          sudo apt-get install luarocks
-          sudo luarocks install --only-deps *.rockspec
-
-      - name: Run Lua unit tests and static code analyzer
-        run: poetry run poe lua-tests
-
-      - name: Run Python tests
-        run: poetry run pytest tests
